seccons commit

+++ src/main/java/com/university/service/Impl/StudentServiceImpl.java
@@ -5,11 +5,12 @@ import com.university.dto.StudentRes;
 import com.university.entity.StudentEntity;
 import com.university.mapper.StudentEntityReqMapper;
 import com.university.repository.StudentRepository;
+import com.university.service.CommonException;
 import com.university.service.StudentService;
-import java.util.HashMap;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -20,18 +21,21 @@ public class StudentServiceImpl implements StudentService {
   StudentRepository studentRepository;
 
   @Override
-  public StudentRes create(StudentReq studentReq) {
-
+  public String create(StudentReq studentReq) {
+    StudentEntity studentEntityCheck = studentRepository.findByCode(studentReq.getCode());
+    getException(studentReq.getCode(), studentReq.getEmail(), studentEntityCheck,
+        studentReq.getPhone());
     StudentEntity studentEntity = StudentEntityReqMapper.INSTANCE.reqToEntity(studentReq);
 
-    StudentEntity studentEntitySave = studentRepository.save(studentEntity);
-    return StudentEntityReqMapper.INSTANCE.entityToRes(studentEntitySave);
+    studentRepository.save(studentEntity);
+    return "Created!!";
   }
 
   @Override
-  public StudentRes update(String code, StudentReq studentReq) {
+  public String update(String code, StudentReq studentReq) {
     StudentEntity studentEntity = studentRepository.findByCode(code);
-
+    getException(null, studentReq.getEmail(), null,
+        studentReq.getPhone());
     studentEntity.setCode(studentReq.getCode());
     studentEntity.setName(studentReq.getName());
     studentEntity.setEmail(studentReq.getEmail());
@@ -43,23 +47,47 @@ public class StudentServiceImpl implements StudentService {
     studentEntity.setBirthdate(studentReq.getBirthdate());
     studentEntity.setMajorsCode(studentReq.getMajorsCode());
 
-    StudentEntity studentEntitySave = studentRepository.save(studentEntity);
+    studentRepository.save(studentEntity);
 
-    return StudentEntityReqMapper.INSTANCE.entityToRes(studentEntitySave);
+    return "Update Successful!!";
   }
 
   @Override
-  public Map<String, Boolean> delete(String code) {
+  public String delete(String code) {
     StudentEntity studentEntity = studentRepository.findByCode(code);
+    if (code.isEmpty()) {
+      throw new CommonException("Code is not null", HttpStatus.BAD_REQUEST, "202");
+    }
     studentRepository.delete(studentEntity);
-    Map<String, Boolean> respone = new HashMap<>();
-    respone.put("Deleted", Boolean.TRUE);
-    return respone;
+    return "Deleted!!";
   }
 
   @Override
-  public List<StudentEntity> getAll(String code) {
+  public List<StudentRes> getAll(String code) {
     List<StudentEntity> studentEntityList = studentRepository.findByCodeOrNull(code);
-    return studentEntityList;
+    List<StudentRes> listStudent = new ArrayList<>();
+    for (StudentEntity student : studentEntityList){
+      StudentRes studentRes;
+      studentRes = StudentEntityReqMapper.INSTANCE.entityToRes(student);
+      listStudent.add(studentRes);
+    }
+    return listStudent;
+  }
+
+  private void getException(String code, String email, StudentEntity studentEntity, Integer phone) {
+    String emailRegex = "^[A-Za-z0-9]+[A-Za-z0-9]*@[A-Za-z0-9]+(\\.[A-Za-z0-9]+)$";
+    String codeRegex = "^[A-Za]{2}+[0-9]{5}$";
+    if (code.isEmpty()) {
+      throw new CommonException("code is not null", HttpStatus.BAD_REQUEST, "1001");
+    } else if (studentEntity != null) {
+      throw new CommonException("Code already exists", HttpStatus.BAD_REQUEST, "1001");
+    } else if (!email.matches(emailRegex)) {
+      throw new CommonException("Email Wrong Format!", HttpStatus.BAD_REQUEST, "1001");
+    } else if (!code.matches(codeRegex)) {
+      throw new CommonException("Code Wrong Format!", HttpStatus.BAD_REQUEST, "1001");
+    } else if (phone.toString().length() < 10) {
+      throw new CommonException("Phone number with minimum 10 characters!", HttpStatus.BAD_REQUEST,
+          "202");
+    }
   }
 }
